changed to use 64-bit floats and changed formula used, bug fixes
performance took a not insignificant hit due to the kernel changes and switch to double precision. Surprisingly, however, runtime did not increase by a large factor (9ish to 13ish seconds on 32 million path simulations). At the moment I'm unsure if this is due to me not implementing the doubles correctly, or due to the limiting steps being in a different part of the program (global memory access, for example, which has a comparatively high latency as opposed to shared and local). In a 'perfect' program, though, runtime would have increased by a factor of at least ten or twenty - wikipedia says 2.7-3.2 GFLOPS for single precision and a measly 116 GFLOPS for double on the 1060 notebook.